authentications and authorized routes handled

+++ models/users.go
@@ -3,7 +3,7 @@
 package models
 
 type User struct {
-	ID       uint   `json:"id" gorm:"primary_key"`
+	ID       uint64  `json:"id" gorm:"primary_key"`
 	Username string `json:"username"`
 	Password string `json:"password"`
 	Name     string `json:"name"`
